NN-2011 Fix issue with conflicting wiremock servers
